event id had type int in commons instead of long. i noticed this when creating EventServiceTest class and changed the type to long.

+++ commons/src/test/java/commons/EventTest.java
@@ -29,7 +29,7 @@ public class EventTest {
 
     @Test
     public void testEventGetters() {
-        int testID = firstEvent.getId();
+        long testID = firstEvent.getId();
         String testName = firstEvent.getName();
         String testInviteCode = firstEvent.getInviteCode();
         List<Participant> testParticipants = firstEvent.getParticipants();
